refactor: extract RetryableSessionPreservationPresenter.swift from TranscriptionRecoveryPresenters.swift
Splits `RetryableSessionPreservationPresenter` (~67 lines) out of
TranscriptionRecoveryPresenters.swift into its own file. The remaining
file holds only RetryTranscriptionStatusPresenter +
PendingTranscriptionRetryFailureHandler.

Byte-identical move of lines 97-164. Both files import Foundation.

TranscriptionRecoveryPresenters.swift: 164 → 95 lines. Tests: 667.
